+++ components/device_shadow/test/host/stubs/nvs_flash.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "nvs.h"
+esp_err_t nvs_flash_init(void);
+esp_err_t nvs_flash_erase(void);